GUI rewrite is mostly complete.  Removed all display list usage and merged everything down to a single vertex buffer object.  The counter and timer widgets (originally based on separate digit DLs) are currently broken.
git-svn-id: https://s.snth.net/svn/neverball/branches/gles@3540 78b8d119-cf0a-0410-b17c-f493084dd1d7

+++ putt/st_all.c
@@ -49,8 +49,8 @@ static char *number(int i)
 }
 
 static int score_card(const char  *title,
-                      const float *c0,
-                      const float *c1)
+                      const GLubyte *c0,
+                      const GLubyte *c1)
 {
     int id, jd, kd, ld;
